+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Now compatible with Alluka `v0.3.0`.
+
 ## [2.17.4] - 2024-04-29
 ### Changed
 - Bumped the minimum Alluka version to `v0.2.0`.